detection should work on every camera

+++ inc/common/Camera.h
@@ -24,7 +24,7 @@ public:
   Camera(common::SharedMemory&);
   ~Camera();
 
-  void getPosition(int& position);
+  void getPosition(float& position);
   int configure();
   void setTracking();
